backend/requirements.txt
# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pyodbc==5.2.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Email
aiosmtplib==3.0.1

# Environment
python-dotenv==1.0.0

# CORS
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2

# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1
click==8.1.7

